refactor(types): decompose lib.rs into focused platform, session, and naming modules [OPE-380]
Extract Platform enum, SessionKey struct, and sanitize_name into dedicated
modules with clear separation of concerns. All 88 tests migrated to tests.rs,
full workspace compilation and downstream imports preserved via lib.rs re-exports.
